question_mark

+++ question_mark/src/lib.rs
@@ -8,11 +8,11 @@ pub struct Three {
      pub third_layer: Option<Four>,
 }
 pub struct Four {
-    pub fourth_layer: Option<u32>,
+    pub fourth_layer: Option<u16>,
 }
 
 impl One {
-    pub fn get_fourth_layer(self) -> Option<u32> {
+    pub fn get_fourth_layer(self) -> Option<u16> {
         self.first_layer?.second_layer?.third_layer?.fourth_layer
     }
